+++ docs/manual/docs/about/release-notes.md
@@ -4,10 +4,19 @@
 
 ### New commands
 
+**SharePoint Online:**
+
+- [spo homesite set](../cmd/spo/homesite/homesite-set.md) - sets the specified site as the Home Site [#1001](https://github.com/pnp/office365-cli/issues/1001)
+
 **Microsoft Graph:**
 
 - [graph o365group user set](../cmd/graph/o365group/o365group-user-set.md) - updates role of the specified user in the specified Office 365 Group or Microsoft Teams team [#982](https://github.com/pnp/office365-cli/issues/982)
 - [graph report teamsdeviceusageuserdetail](../cmd/graph/report/report-teamsdeviceusageuserdetail.md) - gets information about Microsoft Teams device usage by user [#960](https://github.com/pnp/office365-cli/issues/960)
+- [graph teams funsettings set](../cmd/graph/teams/teams-funsettings-set.md) - updates fun settings of a Microsoft Teams team [#817](https://github.com/pnp/office365-cli/issues/817)
+
+**Office 365:**
+
+- [tenant id get](../cmd/tenant-id-get.md) - gets Office 365 tenant ID for the specified domain [#998](https://github.com/pnp/office365-cli/issues/998)
 
 ### Changes
 
@@ -16,6 +25,7 @@
 - added support for superseding SPFx project upgrade findings [#970](https://github.com/pnp/office365-cli/issues/970)
 - added support for package managers [#617](https://github.com/pnp/office365-cli/issues/617)
 - extended 'spo page set' with support for promoting as template [#978](https://github.com/pnp/office365-cli/issues/978)
+- extended 'spo page add' with support for promoting as template [#977](https://github.com/pnp/office365-cli/issues/977)
 
 ## [v1.21.0](https://github.com/pnp/office365-cli/releases/tag/v1.21.0)